+++ app/instance.py
@@ -161,7 +161,7 @@ def create_app():
         from app.api.custom.orders import order_blueprint
         from app.api.custom.invoices import event_blueprint
         from app.api.custom.calendars import calendar_routes
-        from app.api.custom.tickets import tickets_routes
+        from app.api.tickets import tickets_routes
         from app.api.custom.role_invites import role_invites_routes
         from app.api.video_stream import streams_routes
         from app.api.events import events_blueprint
